third_party/blink/web_tests/virtual/controls-refresh/calendar-picker/date-picker-year-view-resize.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../fast/forms/resources/common.js"></script>
<script src="../../../fast/forms/resources/picker-common.js"></script>
<script src="../../../fast/forms/calendar-picker/resources/calendar-picker-common.js"></script>
</head>
<body>

<input type=date id=date value="2019-09-10">

<script>
let t = async_test('Test year list view size');

function test1() {
  var dateViewHeight = popupWindow.global.picker.height();
  clickMonthPopupButton();
  if (popupWindow.global.params.isFormControlsRefreshEnabled) {
    assert_equals(dateViewHeight, popupWindow.global.picker.height());
  } else {
    assert_not_equals(dateViewHeight, popupWindow.global.picker.height());
  }
}

t.step(() => {
  openPicker(document.getElementById('date'), t.step_func_done(test1));
});
</script>
</body>
</html>
